net: move stack into lib.rs
